Add classes for semigroups and monoids
After realizing that we are likely to want semigroup and monoid
for types that don't quite fit into Group (e.g. Naturals), we added
Semigroup and Monoid classes. We have split the Semigroup and Monoid
into two classes each (additive and multiplicative), so we can easily
replicate Num's functionality when it makes sense (e.g. Naturals form
a multiplicative semigroup but not a ring).