Remove smart data dependency
We will instead have smart-data depend on this package. This arguably
makes more sense: smart-data benefits from algebra's instances more than
the reverse. The reason we had algebra depend on smart-data in the first
place was for the NonZero type. Instead we will include that type here and
reverse the dependency.