Remove refined dependency
The primary intention of simple-algebra is to provide a better interface
than Num. Refined was added to give instances to Refined types, but this
was always an uneasy alliance as in a future with dependent types, we
have no need for refined numeric types (and the dependency is not
light).

Given that we have not found much use for the refined instances found
here, (nb. refined is an excellent library and useful _on its own_), we
are removing it. This will:

- Simplify the API and tests
- Make our dependency footprint very light (only dep is base and
  template-haskell)
- Leaves us open to a dependent future.