Add type family constraints to all operators
We will want type families for all operators for future dependent
invariants (e.g. Natural can be an AGroup if, for x .-. y, y includes a
proof that y <= x).

Remove ginv as it is misleading in a future world where x may not have an
additive inverse in general (but nonetheless we can implement y - x sensibly).
This is the same reason we do not have a multInv :: MGroup a => a -> a.