Remove type family constraints
This is sad. The type families offered us a way to allow each instance
to choose its own constraints. On the one hand, 'normal' types like Int
can choose a NonZero constraint for division whereas a 'Positive a'
wrapper can choose itself, since it needs no additional constraint.
On the other hand, these constraints offer the ability to play nicely
with dependent types in the future e.g. we could have an additive
semigroup instance for a bounded type, with the constraint ensuring
that the total sum is less than the bounds.

Unfortunately, there are complications. First of all, it complicates how
we write instances for wrapper types. For example, for our Positive
type, we can write instances like:
  - instance ASemigroup a => ASemigroup (Positive a)
  - instance Num a => ASemigroup (Positive a)
  - instance ASemigroup (Positive Int) ...

These tradeoffs between 1 and 2 involve whether we want to use the
underlying type's constraint, which is not always clear in general.

More seriously, we are totally defeated by existentials. For example,
if we have a GADT like

  data Bar where
    MkBar :: Foo a -> Bar

with Foo having an ASemigroup instance, we may want to write an
instance ASemigroup Bar. Even if we are able to guarantee that the
two Bars wrap the same Foo a, we are unable to write:

    type AddConstraint Bar = forall a. AddConstraint (Foo a)

because we cannot currently include foralls in type families. One may
wonder if this is too contrived, but this was part of the original
motivation for this library; We created a type for bytes that included
a unit for the size to prevent unit errors. Furthermore, we want to
allow reading byte sizes at runtime thus requiring existentially
quantifying the size. Num is convenient but has a number of problems,
hence this library where our Bytes (and others) can opt into exactly
the operations that make sense.

We thus opt to dispense with the type family constraints for now. Yes
it is unfortunate that types like Positive will have a redundant
division constraint, but that is minor and not the norm. The primary
motivation for adding constraints to all operations was anticipating
-XDependentTypes, but this is quite a ways away, and it does not make
sense to handicap ourselves now to _maybe_ make the future easier.
Hopefully by the time we have -XDependentTypes the story will have
improved. In any case, that is a decision best left for then.