Move aabs from AGroup to AMonoid
This is slightly weird as absolute value is trivial for monoids that do
not have inverses. But this _is_ useful (i.e. we want to write code
over some AMonoid and ensure a value is "non-negative"), and it does
not appear to have any downsides.

Strictly speaking, we could use the same logic to move aabs to
ASemigroup. However, it has laws formulated in terms of zero so until
we see a compelling reason to move this to ASemigroup, we will keep it
on AMonoid.

We also remove the abs test for Int. The non-negative law does not hold
for bounded types (e.g. abs (minBound :: Int)).