Replace abs with new Normed typeclass
It feels wrong to pollute the ASemigroup typeclass with abs as the
algebraic notion really requires a full group. Nevertheless it is
useful to enforce a 'must be positive' constraint, and this exists
for types that cannot be monoids (e.g. positives). Thus we split
this functionality into its own typeclass so we can retain this
functionality while not making the algebraic classes messier.