Add MSemiSpace and MSpace
Unfortunately, these classes do not appear to have direct analogues
in abstract algebra. Nevertheless, it is useful to have types that
support 'scala multiplication/division' even if they do not have,
say, an additive monoid (which would otherwise be a Semimodule).
For instance, Q^{2+} -- positive rational numbers -- supports division
but is not an additive monoid, thus ruling out a Semimodule.