Move Literal.X -> Convert.X
Move Literal modules to Convert, to better advertize that these
classes are not just about literals i.e. From/ToReal.

Moreover, we move the implementations into the same module so that we
can provide mutual constraints, as there are clear relationships
between the classes, assuming instances are well-behaved (Natural
negative issues notwithstanding).

Finally, we remove the integral instances for FromRational, since
the instances necessarily make an arbitrary choice for how to handle
non-integers, which we'd rather not make ourselves.